Add Cinemachine to player and create confiner
Cinemachine is a unity camera extension that can be found in the packagemanager. When creating a new camera object from Cinemachine, it takes control over the main camera. This commit adds Cinemachine to our project, and implements the game object "CM vcam1" to track the player within defined container game object "CameraContainer"